refactor: Move checkContractWalletSignature function to a separate file
